pyproject.toml
[project]
name = "egile-agent-scheduler"
version = "0.1.0"
description = "Scheduler for automating Egile agent tasks and team workflows"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "agno>=2.3.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "croniter>=2.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
all = [
    "egile-agent-core @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-core",
    "egile-agent-hub @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-hub",
    "egile-agent-investment @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-investment",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
agent-scheduler = "egile_agent_scheduler.cli:main"
agent-scheduler-daemon = "egile_agent_scheduler.daemon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/egile_agent_scheduler"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
